Same change as -release.
 - Jason


git-svn-id: svn+ssh://localhost/home/epiphani/home/dalrepo/bahamut/trunk@200 57140981-5af7-11db-9af8-080020d9fdd4

+++ src/s_serv.c
@@ -2716,7 +2716,7 @@ m_htm(aClient *cptr, aClient *sptr, int parc, char *parv[])
    char       *command;
 
    extern int  lifesux, LRV, LCF, noisy_htm;	/* in ircd.c */
-   extern float currlife;
+   extern int currlife;
 
    if (!MyClient(sptr) || !IsOper(sptr)) 
    {
